backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.2
openpyxl==3.1.5
pyarrow==17.0.0
openai==1.58.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2

